requirements.txt
streamlit
scikit-learn
numpy
joblib
Pillow